day40 과제완료
