+++ 14th_homework/maincamp/daily-homework/src/app/boards/new/page.tsx
@@ -1,22 +1,12 @@
 'use client';
 
+import { useLoginCheck } from '@/commons/hooks/08-05-use-login-check';
 import BoardsWriteAdvanced from '@/components/boards-write';
-import { useRouter } from 'next/navigation';
-import { useEffect, useRef } from 'react';
+// import { useRouter } from 'next/navigation';
+// import { useEffect, useRef } from 'react';
 
-export default function BoardComponentNewPage() {
-  const router = useRouter();
-  const hasChecked = useRef(false);
-
-  useEffect(() => {
-    if (hasChecked.current) return;
-    hasChecked.current = true;
-
-    if (localStorage.getItem('accessToken') === null) {
-      alert('로그인 후 이용 가능합니다');
-      router.push('/');
-    }
-  }, []);
-
+const BoardNewPage = () =>{
   return <BoardsWriteAdvanced isEdit={false} />;
 }
+
+export default useLoginCheck(BoardNewPage);
